Made some modifications to get the xml-parser working for empty element tags (e.g. <foo/>) and to parse bare attributes (e.g. name='name').
Had to move many productions around as Fun-Dev was complaining often and
loudly about unbound references for methods defined after they were used
in a \with-meta-syntax expansion.

main now takes an argument (the xml to parse) and parses the complete
argument (e.g. multiple tags are parsed).

Still a great deal of work to do, but this is a start that has some
working code.