Made the xml-parser a library instead of an executable.  Created test/xml-test to exercise the library.  It is executed with from the command line thus:
$ ./xml-test "<foo/><bar><baz/>"

And it will show the tags as parsed.

Still severely limitted (see my last CVS entry) -- does not parse text,
does not parse close tags, major pukage or lockage everywhere ...

BUT!

The library and the test executable do work, so this small refactoring
(from an xml-parser executable to an xml-parser library) succeeded.